+++ src/app/components/animal/animal-list/animal-list.component.ts
@@ -1,7 +1,9 @@
 import { Component, OnInit, ViewChild } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
 import { MatPaginator } from '@angular/material/paginator';
 import { MatTableDataSource } from '@angular/material/table';
 import { ToastrService } from 'ngx-toastr';
+import { ConfirmDialogComponent } from 'src/app/confirm-dialog/confirm-dialog.component';
 import { Animal } from 'src/app/models/animal';
 import { AnimalService } from 'src/app/services/animal.service';
 
@@ -17,7 +19,7 @@ export class AnimalListComponent implements OnInit {
   displayedColumns: string[] = ['codigo', 'nome', 'tutor', 'especie', 'pelagem', 'sexo', 'acoes'];
   dataSource = new MatTableDataSource<Animal>(this.ELEMENT_DATA);
 
-  constructor(private service : AnimalService, private toast: ToastrService) { }
+  constructor(private service : AnimalService, private toast: ToastrService, public dialog: MatDialog) { }
 
   ngOnInit(): void {
     this.findAll();
@@ -40,17 +42,23 @@ export class AnimalListComponent implements OnInit {
     })
   }
   excluir(animal : Animal){
-    this.service.delete(animal.idAnimal).subscribe(respota =>{
-      this.toast.success('Animal excluído com sucesso', 'Exclusão')
-      this.findAll();
-    }, ex =>{
-      if(ex.error.errors) {
-        ex.error.errors.forEach(element => {
-          this.toast.error(element.message);
-        });
-      } else {
-        this.toast.error(ex.error.message);
+    const dialogRef = this.dialog.open(ConfirmDialogComponent);
+
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        this.service.delete(animal.idAnimal).subscribe(respota =>{
+          this.toast.success('Animal excluído com sucesso', 'Exclusão')
+          this.findAll();
+        }, ex =>{
+          if(ex.error.errors) {
+            ex.error.errors.forEach(element => {
+              this.toast.error(element.message);
+            });
+          } else {
+            this.toast.error(ex.error.message);
+          }
+        })
       }
-    })
+    });
   }
 }
